functions: add ros_package_exists and ros_workspace
The one is a version of ros_package_assert without a print_fail, just to
detect whether the plugin exists. ros_workspace gives the directory to
the first workspace found. Note that this is not what "roscd" takes you
to, but rather it's the parent directory that has .catkin_workspace. For
me that's where roscd should me take you in the first place.