patches: fawkes-robotino patches for old agent and Gazebo fix
Gazebo seems to have a race condition when loading the "gazebo" binary
with a complex world -- the screen of the GUI will stay black even
though the simulation is fine. There is a workaround separating the
calls to gzserver and gzclient to mitigate the issue.

One rule was missing in the old agent to actually decide for some jobs
to do rather than just wait.